Servant.Auth.JWT

Documentation

class FromJWT a where

How to decode data from a JWT.

The default implementation assumes the data is stored in the unregistered dat claim, and uses the FromJSON instance to decode value from there.

Minimal complete definition

Nothing

Methods

decodeJWT :: ClaimsSet -> Either Text a

default decodeJWT :: FromJSON a => ClaimsSet -> Either Text a

class ToJWT a where

How to encode data from a JWT.

The default implementation stores data in the unregistered dat claim, and uses the type's ToJSON instance to encode the data.

Minimal complete definition

Nothing

Methods

encodeJWT :: a -> ClaimsSet

default encodeJWT :: ToJSON a => a -> ClaimsSet